busman/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via django
django==4.2.7
    # via
    #   -r requirements.in
    #   django-netfields
django-netfields==1.3.1
    # via -r requirements.in
netaddr==0.9.0
    # via django-netfields
six==1.16.0
    # via django-netfields
sqlparse==0.4.4
    # via django
